+++ scripts/macros/2024/classFeatures/cleric/searUndead.js
@@ -1,0 +1,5 @@
+export let searUndead = {
+    name: 'Sear Undead',
+    version: '1.2.12',
+    rules: 'modern'
+};