+++ scripts/macros/2024/classFeatures/cleric/thaumaturge.js
@@ -1,0 +1,5 @@
+export let thaumaturge = {
+    name: 'Divine Order: Thaumaturge',
+    version: '1.2.12',
+    rules: 'modern'
+};